Исправление замечаний для Preview и Question
