+++ GraphLabs.Dal.Ef/packages.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="EntityFramework" version="6.0.1" targetFramework="net45" />
-  <package id="EntityFramework.ru" version="5.0.0" targetFramework="net45" />
   <package id="GraphLabs.Common.Utils" version="1.1.1" targetFramework="net45" />
   <package id="log4net" version="2.0.0" targetFramework="net45" />
   <package id="Unity" version="3.0.1304.1" targetFramework="net45" />
